src/main/java/fr/dwightstudio/jarmemu/asm/inst/STRExecutor.java
package fr.dwightstudio.jarmemu.asm.inst;

import fr.dwightstudio.jarmemu.asm.DataMode;
import fr.dwightstudio.jarmemu.asm.UpdateMode;
import fr.dwightstudio.jarmemu.sim.parse.args.AddressParser;
import fr.dwightstudio.jarmemu.sim.parse.args.ShiftParser;
import fr.dwightstudio.jarmemu.sim.obj.Register;
import fr.dwightstudio.jarmemu.sim.obj.StateContainer;

public class STRExecutor implements InstructionExecutor<Register, AddressParser.UpdatableInteger, Integer, ShiftParser.ShiftFunction> {
    @Override
    public void execute(StateContainer stateContainer, boolean updateFlags, DataMode dataMode, UpdateMode updateMode, Register arg1, AddressParser.UpdatableInteger arg2, Integer arg3, ShiftParser.ShiftFunction arg4) {
        int i1 = arg4.apply(arg3);

        switch (dataMode){
            case null -> stateContainer.memory.putWord(arg2.toInt() + i1, arg1.getData());
            case HALF_WORD -> stateContainer.memory.putHalf(arg2.toInt() + i1, (short) arg1.getData());
            case BYTE -> stateContainer.memory.putByte(arg2.toInt() + i1, (byte) arg1.getData());
        }

        arg2.update();
    }
}
